// apps/web/src/lib/generated/models/ComparisonSetVersionRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ComparisonSetVersionRead = {
  changed_by_user_id: string;
  created_at: string;
  id: string;
  snapshot_json: Record<string, any>;
  version: number;
};